requirements.txt
fastapi
uvicorn
python-pptx
diffusers
transformers
accelerate
bitsandbytes
sentence-transformers
faiss-cpu
pillow
